Rename package from @ottabase/upload to @ottabase/ottaupload
Renamed the package to follow the naming convention of other custom
packages in the monorepo (like ottaeditor, ottaorm, ottarenderer).

Changes:
- Renamed packages/upload to packages/ottaupload
- Updated package name in package.json
- Updated all imports in cloudflare-worker.ts
- Updated all imports in demo page
- Updated package dependency in TanStack app
- Updated README and documentation
